Update iOS and Android code library for any files like audio,image etc
