+++ android/src/main/java/com/azureblobstorage/EAzureBlobStoragePackage.java
@@ -15,7 +15,7 @@ public class EAzureBlobStoragePackage implements ReactPackage {
     @Override
     public List<NativeModule> createNativeModules(ReactApplicationContext reactContext) {
             List<NativeModule> modules = new ArrayList<>();
-            modules.add(new EAzureBlobStorageImage(reactContext));
+            modules.add(new EAzureBlobStorageFile(reactContext));
         return modules;
     }
