+++ ios/EAzureBlobStoragFile.h
@@ -1,0 +1,4 @@
+#import <React/RCTBridgeModule.h>
+
+@interface EAzureBlobStoragFile : NSObject <RCTBridgeModule>
+@end
