--- ios/EAzureBlobStorageImage.h
@@ -1,4 +1,0 @@
-#import <React/RCTBridgeModule.h>
-
-@interface EAzureBlobStorageImage : NSObject <RCTBridgeModule>
-@end
